Move the self-test under spk deployment validate
